+++ src/ScottPlot5/ScottPlot5/Plottables/Heatmap.cs
@@ -4,32 +4,96 @@ public class Heatmap : IPlottable, IHasColorAxis
 {
     public bool IsVisible { get; set; } = true;
     public IAxes Axes { get; set; } = new Axes();
-    public IColormap Colormap { get; set; } = new Colormaps.Viridis();
+    private IColormap _colormap { get; set; } = new Colormaps.Viridis();
+    public IColormap Colormap
+    {
+        get { return _colormap; }
+        set
+        {
+            _colormap = value;
+            Update();
+        }
+    }
 
     /// <summary>
     /// Indicates position of the data point relative to the rectangle used to represent it.
     /// An alignment of upper right means the rectangle will appear to the lower left of the point itself.
     /// </summary>
-    public Alignment CellAlignment { get; set; } = Alignment.MiddleCenter;
+    private Alignment _cellAlignment { get; set; } = Alignment.MiddleCenter;
+    public Alignment CellAlignment
+    {
+        get { return _cellAlignment; }
+        set
+        {
+            _cellAlignment = value;
+            Update();
+        }
+    }
 
     /// <summary>
     /// If defined, the this rectangle sets the axis boundaries of heatmap data.
     /// Note that the actual heatmap area is 1 cell larger than this rectangle.
     /// </summary>
-    public CoordinateRect? Extent { get; set; }
+    private CoordinateRect? _extent;
+    public CoordinateRect? Extent
+    {
+        get { return _extent; }
+        set
+        {
+            _extent = value;
+            Update();
+        }
+    }
 
     /// <summary>
     /// This variable controls whether row 0 of the 2D source array is the top or bottom of the heatmap.
     /// When set to false (default), row 0 is the top of the heatmap.
     /// When set to true, row 0 of the source data will be displayed at the bottom.
     /// </summary>
-    public bool FlipVertically { get; set; } = false;
+    private bool _flipRows = false;
+    public bool FlipVertically
+    {
+        get { return _flipRows; }
+        set
+        {
+            _flipRows = value;
+            Update();
+        }
+    }
+
+    /// <summary>
+    /// This variable controls whether the first sample in each column of the 2D source array is the left or right of the heatmap.
+    /// When set to false (default), sample 0 is the left of the heatmap.
+    /// When set to true, sample 0 of the source data will be displayed at the right.
+    /// </summary>
+    private bool _flipColumns = false;
+    public bool FlipHorizontally
+    {
+        get { return _flipColumns; }
+        set
+        {
+            _flipColumns = value;
+            Update();
+        }
+    }
+
+    public bool FlipColumns { get => FlipHorizontally; set => FlipHorizontally = value; }
+    public bool FlipRows { get => FlipVertically; set => FlipVertically = value; }
 
     /// <summary>
     /// If true, pixels in the final image will be interpolated to give the heatmap a smooth appearance.
     /// If false, the heatmap will appear as individual rectangles with sharp edges.
     /// </summary>
-    public bool Smooth { get; set; } = false;
+    private bool _smooth { get; set; } = false;
+    public bool Smooth
+    {
+        get { return _smooth; }
+        set
+        {
+            _smooth = value;
+            Update();
+        }
+    }
 
     /// <summary>
     /// Actual extent of the heatmap bitmap after alignment has been applied
@@ -38,9 +102,9 @@ public class Heatmap : IPlottable, IHasColorAxis
     {
         get
         {
-            double x = CellWidth * CellAlignment.HorizontalFraction();
-            double y = CellHeight * CellAlignment.VerticalFraction();
-            Coordinates cellOffset = new(-x, -y);
+            double xOffset = Math.Abs(CellWidth) * CellAlignment.HorizontalFraction();
+            double yOffset = Math.Abs(CellHeight) * CellAlignment.VerticalFraction();
+            Coordinates cellOffset = new(-xOffset, -yOffset);
             return ExtentOrDefault.WithTranslation(cellOffset);
         }
     }
@@ -55,7 +119,18 @@ public class Heatmap : IPlottable, IHasColorAxis
         get
         {
             if (Extent.HasValue)
-                return Extent.Value;
+            {
+                var extent = Extent.Value;
+                //user will provide the extends to the data. The image will be one cell wider and taller so we need to add that on (it is being added on in teh default case).
+                double cellwidth = extent.Width / (Intensities.GetLength(1) - 1);
+                double cellheight = extent.Height / (Intensities.GetLength(0) - 1);
+                if (extent.Left < extent.Right) extent.Right += cellwidth;
+                if (extent.Left > extent.Right) extent.Left -= cellwidth; //cellwidth will be negative if extent is flipped
+                if (extent.Bottom < extent.Top) extent.Top += cellheight;
+                if (extent.Bottom > extent.Top) extent.Bottom -= cellheight; //cellheight will be negative if extent is inverted
+
+                return extent;
+            }
 
             return new CoordinateRect(
                 left: 0,
@@ -116,12 +191,18 @@ public class Heatmap : IPlottable, IHasColorAxis
     {
         Range range = GetRange();
         uint[] argb = new uint[Intensities.Length];
+
+        // the XOR here disables flipping when the flip property and the extent is inverted.
+        bool FlipY = FlipVertically ^ ExtentOrDefault.IsInvertedY;
+        bool FlipX = FlipHorizontally ^ ExtentOrDefault.IsInvertedX;
+
         for (int y = 0; y < Height; y++)
         {
-            int rowOffset = FlipVertically ? (Height - 1 - y) * Width : y * Width;
+            int rowOffset = FlipY ? (Height - 1 - y) * Width : y * Width;
             for (int x = 0; x < Width; x++)
             {
-                argb[rowOffset + x] = Colormap.GetColor(Intensities[y, x], range).ARGB;
+                int xIndex = FlipX ? (Width - 1 - x) : x;
+                argb[rowOffset + x] = Colormap.GetColor(Intensities[y, xIndex], range).ARGB;
             }
         }